# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Locação API",
    "version": "1.0.0",
    "description": "API para gerenciamento de locações e brinquedos."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor Local"
    }
  ],
  "paths": {
    "/brinquedos": {
      "get": {
        "summary": "Listar todos os brinquedos",
        "operationId": "getBrinquedos",
        "responses": {
          "200": {
            "description": "Lista de brinquedos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BrinquedoResponseComTipoDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar um novo brinquedo",
        "operationId": "createBrinquedo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBrinquedoDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Brinquedo criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrinquedoResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/brinquedos/{cod}": {
      "parameters": [
        {
          "name": "cod",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid",
            "example": "8e07cb00-089a-4e0f-b366-3889bf2c61db"
          }
        }
      ],
      "get": {
        "summary": "Obter um brinquedo pelo código",
        "operationId": "getBrinquedoById",
        "responses": {
          "200": {
            "description": "Brinquedo encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrinquedoResponseComTipoDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar um brinquedo",
        "operationId": "updateBrinquedo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBrinquedoDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Brinquedo atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrinquedoResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Excluir um brinquedo",
        "operationId": "deleteBrinquedo",
        "responses": {
          "204": {
            "description": "Brinquedo removido"
          }
        }
      }
    },
    "/brinquedosLocados": {
      "get": {
        "summary": "Listar todos os brinquedos locados",
        "operationId": "getBrinquedosLocados",
        "responses": {
          "200": {
            "description": "Lista de brinquedos locados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BrinquedoLocadoResponseDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar um novo brinquedo locado",
        "operationId": "createBrinquedoLocado",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBrinquedoLocadoDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Brinquedo locado criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrinquedoLocadoResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/brinquedosLocados/{cod}": {
      "parameters": [
        {
          "name": "cod",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid",
            "example": "8e07cb00-089a-4e0f-b366-3889bf2c61db"
          }
        }
      ],
      "get": {
        "summary": "Obter um brinquedo locado pelo código",
        "operationId": "getBrinquedoLocadoById",
        "responses": {
          "200": {
            "description": "Brinquedo locado encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrinquedoLocadoResponseDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar um brinquedo locado",
        "operationId": "updateBrinquedoLocado",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBrinquedoLocadoDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Brinquedo locado atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrinquedoLocadoResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Excluir um brinquedo locado",
        "operationId": "deleteBrinquedoLocado",
        "responses": {
          "204": {
            "description": "Brinquedo locado removido"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateBrinquedoDTO": {
        "type": "object",
        "required": [
          "nome",
          "tipo_brinquedo",
          "marca",
          "data_aquisicao",
          "valor_locacao"
        ],
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255,
            "example": "Brinquedo X"
          },
          "tipo_brinquedo": {
            "type": "string",
            "format": "uuid",
            "example": "uuid-tipo"
          },
          "marca": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255,
            "example": "Marca Y"
          },
          "data_aquisicao": {
            "type": "string",
            "format": "date",
            "example": "2023-12-31"
          },
          "valor_locacao": {
            "type": "number",
            "minimum": 0,
            "example": 100.5
          }
        },
        "additionalProperties": false
      },
      "UpdateBrinquedoDTO": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "tipo_brinquedo": {
            "type": "string",
            "format": "uuid"
          },
          "marca": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "data_aquisicao": {
            "type": "string",
            "format": "date"
          },
          "valor_locacao": {
            "type": "number",
            "minimum": 0
          },
          "ativo": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "BrinquedoResponseDTO": {
        "type": "object",
        "properties": {
          "cod": {
            "type": "string",
            "format": "uuid",
            "example": "8e07cb00-089a-4e0f-b366-3889bf2c61db"
          },
          "nome": {
            "type": "string",
            "example": "Brinquedo X"
          },
          "tipo_brinquedo": {
            "type": "string",
            "format": "uuid",
            "example": "8e07cb00-089a-4e0f-b366-3889bf2c61db"
          },
          "marca": {
            "type": "string",
            "example": "Marca Y"
          },
          "data_aquisicao": {
            "type": "string",
            "format": "date",
            "example": "2023-12-31"
          },
          "valor_locacao": {
            "type": "number",
            "example": 100.5
          },
          "ativo": {
            "type": "boolean"
          }
        }
      },
      "BrinquedoResponseComTipoDTO": {
        "type": "object",
        "properties": {
          "cod": {
            "type": "string",
            "format": "uuid",
            "example": "8e07cb00-089a-4e0f-b366-3889bf2c61db"
          },
          "nome": {
            "type": "string",
            "example": "Brinquedo X"
          },
          "tipo_brinquedo": {
            "type": "string",
            "format": "uuid",
            "example": "uuid-tipo"
          },
          "marca": {
            "type": "string",
            "example": "Marca Y"
          },
          "data_aquisicao": {
            "type": "string",
            "format": "date",
            "example": "2023-12-31"
          },
          "valor_locacao": {
            "type": "number",
            "example": 100.5
          },
          "tipoBrinquedo": {
            "$ref": "#/components/schemas/TipoBrinquedoResponseDTO"
          },
          "ativo": {
            "type": "boolean"
          }
        }
      },
      "CreateBrinquedoLocadoDTO": {
        "type": "object",
        "required": ["cod_locacao", "data_devolucao", "cod_brinquedo"],
        "properties": {
          "cod_locacao": {
            "type": "string",
            "format": "uuid",
            "example": "8e07cb00-089a-4e0f-b366-3889bf2c61db"
          },
          "cod_brinquedo": {
            "type": "string",
            "format": "uuid",
            "example": "8e07cb00-089a-4e0f-b366-3889bf2c61db"
          }
        },
        "additionalProperties": false
      },
      "UpdateBrinquedoLocadoDTO": {
        "type": "object",
        "properties": {
          "cod_locacao": {
            "type": "string",
            "format": "uuid",
            "example": "8e07cb00-089a-4e0f-b366-3889bf2c61db"
          },
          "cod_brinquedo": {
            "type": "string",
            "format": "uuid",
            "example": "8e07cb00-089a-4e0f-b366-3889bf2c61db"
          },
          "ativo": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "BrinquedoLocadoResponseDTO": {
        "type": "object",
        "properties": {
          "cod": {
            "type": "string",
            "format": "uuid",
            "example": "8e07cb00-089a-4e0f-b366-3889bf2c61db"
          },
          "valor_unitario": {
            "type": "number",
            "example": 50.0
          },
          "nome": {
            "type": "string",
            "example": "Brinquedo Locado X"
          },
          "cod_locacao": {
            "type": "string",
            "format": "uuid",
            "example": "8e07cb00-089a-4e0f-b366-3889bf2c61db"
          },
          "data_devolucao": {
            "type": "string",
            "format": "date",
            "example": "2024-01-01"
          },
          "cod_brinquedo": {
            "type": "string",
            "format": "uuid",
            "example": "8e07cb00-089a-4e0f-b366-3889bf2c61db"
          },
          "ativo": {
            "type": "boolean"
          }
        }
      },
...
      "TipoBrinquedoResponseDTO": {
        "type": "object",
        "properties": {
          "cod": {
            "type": "string",
            "format": "uuid",
            "example": "8e07cb00-089a-4e0f-b366-3889bf2c61db"
          },
          "nome": {
            "type": "string"
          },
          "ativo": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}